linux-yocto/5.15: fix -standard kernel build issue
Integrating the following commit(s) to linux-yocto/5.15:

    4eba9348d3e2 Revert "Revert "fbdev: Hot-unplug firmware fb devices on forced removal""

The revert of commit [fbdev: Hot-unplug firmware fb devices on forced
removal] was done to fix powerpc fbdev issues. Upstream went in a
different direction, which means that our fbdev routines have conflicts
with -stable updates.

The fix for the fbdev is in -stable, so we drob our reverted commit,
such that 5.15 builds again, and fbdev is functional.

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>